+++ client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "client2",
+  "name": "client",
   "private": true,
   "version": "0.0.0",
   "type": "module",
