+++ scripts/Caddyfile
@@ -8,19 +8,8 @@ localhost {
 		path /signup_callback
 	}
 
-	# handle @apipath {
-	# 	reverse_proxy {
-	# 		to rss_server:8080
-	# 	}
-	# }
-
-    # handle {
-    #     reverse_proxy {
-    #         to rss_client
-    #     }
-    # }
-
-  # Useful for local development      
+   # Local development
+   # --------------------------------------
    reverse_proxy host.docker.internal:8081 {
       header_up X-Real-IP {remote_host}
    }
@@ -30,4 +19,19 @@ localhost {
 			header_up X-Real-IP {remote_host}
 		}
 	}
-}   
+
+
+    # Dockers reverse proxies
+    # --------------------------------------
+	# handle @apipath {
+	# 	reverse_proxy {
+	# 		to rss_server:8080
+	# 	}
+	# }
+    #
+    # handle {
+    #     reverse_proxy {
+    #         to rss_client
+    #     }
+    # }
+}
